Fix : Restructuring files and refactoring ToolTest
