+++ dev/io.openliberty.mcp.internal_fat/fat/src/io/openliberty/mcp/internal/fat/tool/businessExceptionApp/UnwrappedBusinessException.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package io.openliberty.mcp.internal.fat.tool.basicToolApp.BusinessException;
+package io.openliberty.mcp.internal.fat.tool.businessExceptionApp;
 
 /**
  *
